Initial version of the algorithm.ilp.neos plugin
It provides an "experimental" interface to the CPLEX solver
(experimental, because it is not unit-tested yet). However,
it can be used to solve the drwahnsinn example. Also, the
way we send the AMPL data to the neos server is not very
efficient (see code comments). All in all, this plugin needs
more love.