Add testcases for the sequencePart concept
In addition to explicitly listing the named parts in a tuple descriptor,
it is now possible to specify a single so-called sequencePart. If the
sequencePart is set, then the to be defined tuple type consists of
exactly one named part, whose name is the value of the specified
sequencePart. Its type is Sequence(OclAny). Informally, the semantics
of such a tuple is as follows (assuming the value of the sequencePart
is "foo"):

Tuple {
	foo = Sequence {
		obj1,
		obj2,
		obj3,
		obj4,
		...
		objN,
		objN+1
	}
}

obj1 is related to obj2 and obj3 is related to obj4 and ... and objN
is related to objN+1 (note that N+1 is even).

Using a sequence instead of a fixed-length (, predefined) OCL tuple
is handy, if one needs to reason about the relation of multiple
elements, where the exact number of elements is not known at the
specification time (for instance, the number could vary from model
to model or even from model element to model element etc.). A potential
(MUML) use case for this might be connector allocation (where a connector
is related (or "allocated") to a "route" (usually, not all routes have
the same length) (a route is no model element; essentially, a route
is made up of several hardware connectors => "allocating" a connector
to a route means relating a connector to several hardware connectors)).