Speed up Specification2ILP transformation
Using something like "ilp.variables += ..." is a performance nightmare,
because the ilp.variables "list" is copied/iterated several times. For
the details, see
org.eclipse.m2m.internal.qvt.oml.ast.env.QvtOperationalEvaluationEnv.assign.

To cope with this, we wrap such "lists" in qvto Lists (since a qvto List
is mutable and, hence, not copied during such a "+=" operation). Eventually,
we sync the wrapped lists, which results in a single copy. Note that the
current implementation is a bit hacky, because we rely on the fact that
every model element can be casted to an EObject...