Get rid of the unused "storeILPModel" hack
Currently, it is unused and will be implemented in a more appropriate
way (from a conceptual point of view, it does not belong to the
IAllocationComputationStrategy (for instance, not every acs uses an
ILP-based approach)). Unfortunately, this "breaks" the API of our
interfaces... (but since we are the only users... ;) ).