Move unsafe cast from the AllocationComputationOperation to the wizard
Rationale: the AllocationComputationOperation could be reused in some
more general contexts => it should be type safe. Unfortunately, there
is no easy way to get rid of the cast. We could get rid of the generics
in the interfaces, though, but this would be only beneficial in this
particular use case... "All" other use cases (which we admittedly don't
have at the moment) would become more cumbersome without generics.
In short, we do the unsafe cast.