Introduce FixedStructureILPComparator and testcases
Can be used to compare ILPs (and expressions). Note that this only
works for ILPs that have a specific structure. Its main use case is
to compare ILPs that are generated by the Specification2ILP.qvto
transformation (hence the specific structure). Later, it will be
used to test the transformation results in case of an
"unreproducible" Specification2ILP.qvto transformation run (that is,
most of the transformation's sorting code is skipped).

Big fat warning: this code is desperately looking for more love!