Introduce config option to generate a reproducible ILP
By default, the reproducibleILP option is set to "true". That is,
two subsequent runs of the Specification2ILP.qvto transformation will
output the identical ILP models. However, generating a reproducible
ILP is "time consuming", because the current implementation does a
lot of sorting. Hence, by setting the reproducibleILP to "false", one
might gain a speed up.
Note: If we would change the default to "false", we would need to
adapt all existing testcases (so that they explicitly set the
reproducibleILP option to "true") (that's why the current default
is "true").
Note 2: the LocationConstraint.qvto will be adapted in a separate
commit (since we are going to add other speed improvements in one go).